FIX: removed useless stuff
