+++ .env
@@ -6,12 +6,8 @@ VUE_APP_BACKEND_ITEMS=items/
 VUE_APP_BACKEND_VIDEOS=videos/
 VUE_APP_BACKEND_QUESTIONS=questions/
 
-VUE_APP_BACKEND_UPDATE_ENTRY=/entries/update
 
 VUE_APP_BACKEND_EXPERIMENT_ASSIGNMENT=/experiments/get_assignment
-
-VUE_APP_BACKEND_USER_CONFIG=/users/get_config
-VUE_APP_BACKEND_UPDATE_USER_CONFIG=/users/update_config
 VUE_APP_BACKEND_LOGIN_USER=/users/login
 
 VUE_APP_I18N_LOCALE=hi
