+++ src/router/index.js
@@ -2,8 +2,7 @@ import { createRouter, createWebHashHistory } from "vue-router";
 import Home from "@/pages/Home.vue";
 import Test from "@/pages/Test.vue";
 import Editor from "@/pages/Editor.vue";
-// import Player from "@/pages/Player.vue";
-import NewPlayer from "@/pages/NewPlayer.vue";
+import Player from "@/pages/Player.vue";
 import PhoneSignIn from "@/pages/PhoneSignIn";
 
 const routes = [
@@ -26,7 +25,7 @@ const routes = [
   {
     path: "/play/:id",
     name: "Player",
-    component: NewPlayer,
+    component: Player,
     query: {
       src: "",
     },
