+++ src/services/API/User.js
@@ -1,13 +1,6 @@
 import { apiClient } from "@/services/API/RootClient.js";
 
 export default {
-  postUserResponse(userResponse) {
-    return apiClient().post(
-      process.env.VUE_APP_BACKEND_UPDATE_ENTRY,
-      userResponse
-    );
-  },
-
   loginUser(userCreds) {
     return apiClient().post(process.env.VUE_APP_BACKEND_LOGIN_USER, userCreds);
   },
